requirements.txt
flask
python-dotenv
reportlab
pandas
pytest
pypdf
pytest-mock